feat(hvmm): streamline firmware handling and improve error reporting
- Removed the hardcoded default firmware path in favor of a new `ResolveFirmware` function to handle firmware resolution dynamically.
- Updated the `Boot` and `SnapshotRoundTrip` methods to utilize the new firmware resolution logic, enhancing flexibility.
- Improved error handling in the `hvErr` function to ensure only the lower 32 bits of the return code are considered, preventing potential issues with junk data in higher bits.
- Adjusted watchdog timing in the `Boot` function for better guest liveness sampling, ensuring more accurate monitoring of guest state.

This refactor enhances the overall robustness and maintainability of the hvmm module.

+++ internal/hvmm/hvmm.go
@@ -35,7 +35,9 @@ const (
 
 // hvErr turns a non-zero hv_return_t into an error.
 func hvErr(what string, rc int) error {
-	if rc == 0 {
+	// hv_return_t is a 32-bit value; some purego trampolines leave junk in the
+	// upper bits of the returned Go int, so only the low 32 bits are significant.
+	if uint32(rc) == 0 {
 		return nil
 	}
 	return fmt.Errorf("%s: hv_return=0x%08x", what, uint32(rc))
